+++ CHANGELOG.md
@@ -4,9 +4,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## Unreleased
+### Removed
+- Support for basic auth as an option for Gitea, by [@techknowlogick](https://giteahub.com/techknowlogick). [#2721](https://github.com/drone/drone/issues/2721)
+
 ## [1.3.1] - 2019-08-26
 ### Added
-
 - support for the GitHub deployment status API, by [@bradrydzewski](https://github.com/bradrydzewski).
 
 ## [1.3.0] - 2019-08-20