+++ CHANGELOG.md
@@ -4,6 +4,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## Unreleased
+### Added
+- support for configuration conversion plugins (e.g. convert from starlark)
+- moved native jsonnet to a conversion plugin
+
 ## [1.3.1] - 2019-08-26
 ### Added